+++ HISTORY.rst
@@ -10,3 +10,25 @@ History
 ------------------
 
 * Second alpha release on GitHub.
+
+0.3.0 (2021-10-01)
+------------------
+
+* First stable feature-rich release
+
+0.4.0 (2022-02-23)
+------------------
+
+* GUI for MD DaVis added; invoke with md-davis-gui
+* Fixed plotting xvg files without title
+* Changed data type of resi from int to int32 in collate.py
+
+0.4.1 (2022-03-02)
+------------------
+
+* Fixed installation on Linux due to open-source PyMOL 2.4.0
+* Fixing readthedocs documentation compilation
+* Bug Fix: landscape_animation was not working due change of landscapes from array to dict
+* Removed docopt dependency
+* Renamed md_davis to md-davis in documentations
+